+++ src/backend/cpu/simd_x86_256.h
@@ -14,7 +14,7 @@ struct FVec {
 struct IVec {
     IVec() = default;
     IVec(__mm256i _v) : v(_v) {};
-    IVec(long long _v) : v(vdupq_n_s64(_v)) {};
+    IVec(int64_t _v) : v(vdupq_n_s64(_v)) {};
     operator __mm256i() { return v; }
     __mm256i v;
 };
